pyproject.toml
[project]
name = "rag-internvl-poc"
version = "0.1.0"
description = "POC RAG avec InternVL3.5-8B, Postgres pgvector et FTS hybride (BM25-like)"
readme = "README.md"
requires-python = ">=3.11,<3.12"
license = { text = "Apache-2.0" }
authors = [
  { name = "POC", email = "poc@example.com" }
]
dependencies = [
  "psycopg[binary]>=3.2.1",
  "numpy>=1.26",
  "pydantic>=2.7",
  "python-dotenv>=1.0",
  "tqdm>=4.66",
  "pymupdf>=1.24.5",
  "Pillow>=10.4.0",
  "sentence-transformers>=3.0.1",
  "torch>=2.3.0",
  "transformers>=4.52.1",
  "typer>=0.12.3",
]

[project.scripts]
rag-internvl = "rag_internvl_poc.cli:app"

[tool.uv]
# Outils de dev facultatifs
dev-dependencies = [
  "ruff>=0.5.6",
  "mypy>=1.11.0",
]

